chore(crd): move validation functions to validation.go
