+++ pkg/apis/spinnaker/v1alpha2/spinnakerservice_types.go
@@ -4,10 +4,8 @@ import (
 	"crypto/md5"
 	"encoding/hex"
 	"encoding/json"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	"time"
-
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
@@ -151,30 +149,6 @@ type SpinnakerServiceStatus struct {
 	AccountCount int `json:"accountCount,omitempty"`
 }
 
-// GetUpdateHash updates the hash at key `key` and returns the prior copy if one existed
-// LastDeployed should then contain the hash and the time if updateTime is true or if there was no hash
-func (s *SpinnakerServiceStatus) UpdateHashIfNotExist(key, hash string, t time.Time, updateTime bool) HashStatus {
-	if s.LastDeployed == nil {
-		s.LastDeployed = make(map[string]HashStatus)
-	}
-	res := HashStatus{}
-	ld, ok := s.LastDeployed[key]
-	if ok {
-		ld.DeepCopyInto(&res)
-		ld.Hash = hash
-		if updateTime {
-			ld.LastUpdatedAt = metav1.NewTime(t)
-		}
-	} else {
-		ld = HashStatus{
-			Hash:          hash,
-			LastUpdatedAt: metav1.NewTime(t),
-		}
-	}
-	s.LastDeployed[key] = ld
-	return res
-}
-
 // +k8s:openapi-gen=true
 type HashStatus struct {
 	Hash          string      `json:"hash"`
@@ -204,14 +178,6 @@ type SpinnakerValidation struct {
 	CI map[string]ValidationSetting `json:"ci,omitempty"`
 }
 
-func (s *SpinnakerValidation) GetValidationSettings() ValidationSetting {
-	return ValidationSetting{
-		Enabled:          true,
-		FailOnError:      s.FailOnError,
-		FrequencySeconds: s.FrequencySeconds,
-	}
-}
-
 type ValidationSetting struct {
 	// Enable or disable validation, defaults to false
 	Enabled bool `json:"enabled"`
@@ -221,25 +187,6 @@ type ValidationSetting struct {
 	// Number of seconds between each validation
 	// +optional
 	FrequencySeconds intstr.IntOrString `json:"frequencySeconds,omitempty"`
-}
-
-func (v *ValidationSetting) NeedsValidation(lastValid metav1.Time) bool {
-	if lastValid.IsZero() {
-		return true
-	}
-	secs := v.FrequencySeconds.IntValue()
-	if secs == 0 {
-		return false
-	}
-	n := lastValid.Time.Add(time.Duration(secs) * time.Second)
-	return time.Now().After(n)
-}
-
-func (v *ValidationSetting) IsFatal() bool {
-	if v.FailOnError == nil {
-		return true
-	}
-	return *v.FailOnError
 }
 
 func init() {
